# eyewear_node_typescript/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Eyewear API"
  },
  "basePath": "/api/v1",
  "parameters": {
    "Authrization": {
      "in": "header",
      "name": "authorization",
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q",
      "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q"
    },
    "Timezone": {
      "in": "header",
      "name": "timezone",
      "type": "string",
      "example": "Asia/Calcutta",
      "default": "Asia/Calcutta",
      "required": true
    },
    "AcceptLanguage": {
      "in": "header",
      "name": "Accept-Language",
      "type": "string",
      "example": "en/hi",
      "default": "hi",
      "required": true
    }
  },
...
  "paths": {
...
    "/carts": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "get user cart api",
        "description": "get user cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true,
            "example": "616a82d2740e9936893fe7b4",
            "default": "616a82d2740e9936893fe7b4"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartGet"
            }
          }
        }
      }
    },
    "/carts/": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "add cart api",
        "description": "add cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "example": "616a82d2740e9936893fe7b4"
                },
                "product_id": {
                  "type": "string",
                  "example": "616325ececd4223283d4f123"
                },
                "quantity": {
                  "required": true,
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartAdd"
            }
          }
        }
      }
    },
    "/carts/{id}": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "put": {
        "summary": "update cart api",
        "description": "update cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "example": "617442a33fd5225b3c8f2b5a",
            "default": "617442a33fd5225b3c8f2b5a"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "quantity": {
                  "required": true,
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartUpdate"
            }
          }
        }
      }
    },
    "/carts/{id}/": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "delete": {
        "summary": "delete cart api",
        "description": "delete cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "example": "617442a33fd5225b3c8f2b5a",
            "default": "617442a33fd5225b3c8f2b5a"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartDelete"
            }
          }
        }
      }
    },
    "/address": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "get user address api",
        "description": "get user address API",
        "tags": ["Address apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true,
            "example": "616a82d2740e9936893fe7b4",
            "default": "616a82d2740e9936893fe7b4"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/addressGet"
            }
          }
        }
      }
    },
...
    "/address/{id}": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "put": {
        "summary": "add address api",
        "description": "add address API",
        "tags": ["Address apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "example": "61742ec7d3ece53de339d5ec",
            "default": "61742ec7d3ece53de339d5ec"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "61742ec7d3ece53de339d5ec"
                },
                "user_id": {
                  "type": "string",
                  "example": "616a82d2740e9936893fe7b4",
                  "required": true
                },
                "street": {
                  "type": "string",
                  "example": 16,
                  "required": true
                },
                "address_type": {
                  "required": true,
                  "type": "string",
                  "example": "Home"
                },
                "address": {
                  "required": true,
                  "type": "string",
                  "example": "yamuna vihar"
                },
                "landmark": {
                  "required": true,
                  "type": "string",
                  "example": "golden public school"
                },
                "city": {
                  "required": true,
                  "type": "string",
                  "example": "new delhi"
                },
                "state": {
                  "required": true,
                  "type": "string",
                  "example": "delhi"
                },
                "zipcode": {
                  "required": true,
                  "type": "number",
                  "example": 110044
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/addressUpdate"
            }
          }
        }
      }
    },
    "/address/{id}/": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "show address api",
        "description": "show  address API",
        "tags": ["Address apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "example": "61742ec7d3ece53de339d5ec",
            "default": "61742ec7d3ece53de339d5ec"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/addressShow"
            }
          }
        }
      }
    },
    "/address/{id}?type=delete": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "delete": {
        "summary": "delete address api",
        "description": "delete  address API",
        "tags": ["Address apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "example": "6175908b2e8f9b7e4471520f",
            "default": "6175908b2e8f9b7e4471520f"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/addressDelete"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "cartGet": {
      "allOf": [
        {
          "$ref": "#/definitions/cartGet"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "cart retrived successfullly!"
            }
          }
        }
      ]
    },
    "cartAdd": {
      "allOf": [
        {
          "$ref": "#/definitions/cartAdd"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "add to cart successfullly!"
            }
          }
        }
      ]
    },
    "cartUpdate": {
      "allOf": [
        {
          "$ref": "#/definitions/cartUpdate"
        },
        {
          "type": "object"
        },
        {
          "post": null,
          "properties": {
            "message": {
              "type": "string",
              "example": "update cart successfullly!"
            }
          }
        }
      ]
    },
    "cartDelete": {
      "allOf": [
        {
          "$ref": "#/definitions/cartDelete"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "cart deleted successfullly!"
            }
          }
        }
      ]
    },
    "addressGet": {
      "allOf": [
        {
          "$ref": "#/definitions/addressGet"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "address retrived successfullly!"
            }
          }
        }
      ]
    },
...
    "addressUpdate": {
      "allOf": [
        {
          "$ref": "#/definitions/addressUpdate"
        },
        {
          "type": "object"
        },
        {
          "post": null,
          "properties": {
            "message": {
              "type": "string",
              "example": "update address successfullly!"
            }
          }
        }
      ]
    },
    "addressShow": {
      "allOf": [
        {
          "$ref": "#/definitions/addressShow"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "address retrived successfullly!"
            }
          }
        }
      ]
    },
    "addressDelete": {
      "allOf": [
        {
          "$ref": "#/definitions/addressDelete"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "address deleted successfullly!"
            }
          }
        }
      ]
    },
...
  }
}